Move reporting functions to report_util
