+++ src/cbmc/bmc.cpp
@@ -25,6 +25,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <linking/static_lifetime_init.h>
 
 #include <goto-checker/bmc_util.h>
+#include <goto-checker/report_util.h>
 #include <goto-checker/solver_factory.h>
 
 #include "counterexample_beautification.h"
