+++ src/goto-checker/report_util.cpp
@@ -17,6 +17,62 @@ Author: Daniel Kroening, Peter Schrammel
 #include <util/ui_message.h>
 #include <util/xml.h>
 
+void report_success(ui_message_handlert &ui_message_handler)
+{
+  messaget msg(ui_message_handler);
+  msg.result() << "VERIFICATION SUCCESSFUL" << messaget::eom;
+
+  switch(ui_message_handler.get_ui())
+  {
+  case ui_message_handlert::uit::PLAIN:
+    break;
+
+  case ui_message_handlert::uit::XML_UI:
+  {
+    xmlt xml("cprover-status");
+    xml.data = "SUCCESS";
+    msg.result() << xml;
+  }
+  break;
+
+  case ui_message_handlert::uit::JSON_UI:
+  {
+    json_objectt json_result;
+    json_result["cProverStatus"] = json_stringt("success");
+    msg.result() << json_result;
+  }
+  break;
+  }
+}
+
+void report_failure(ui_message_handlert &ui_message_handler)
+{
+  messaget msg(ui_message_handler);
+  msg.result() << "VERIFICATION FAILED" << messaget::eom;
+
+  switch(ui_message_handler.get_ui())
+  {
+  case ui_message_handlert::uit::PLAIN:
+    break;
+
+  case ui_message_handlert::uit::XML_UI:
+  {
+    xmlt xml("cprover-status");
+    xml.data = "FAILURE";
+    msg.result() << xml;
+  }
+  break;
+
+  case ui_message_handlert::uit::JSON_UI:
+  {
+    json_objectt json_result;
+    json_result["cProverStatus"] = json_stringt("failure");
+    msg.result() << json_result;
+  }
+  break;
+  }
+}
+
 void output_properties(
   const propertiest &properties,
   ui_message_handlert &ui_message_handler)
